moved concurrent sending code to SendBulkEmail func
